src/auth-service/requirements.txt
astroid==2.12.13
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
dill==0.3.6
Flask==2.2.2
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
parso==0.8.3
platformdirs==2.5.4
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.6
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
python-json-logger~=3.3.0
